+++ src/components/NavItems.jsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-let purple = '#9580ff'
+let purple = '#bd93f9'
 
 export const NavItems = styled.div`
     display: flex;
@@ -9,14 +9,14 @@ export const NavItems = styled.div`
     right: 50px;
     top: -0.5rem;
     width: 23rem;
-    
+
     p {
         font-size: 2rem;
         font-weight: 400;
         margin: 28px 10px 0 0;
 
         a {
-            background: -webkit-linear-gradient(90deg, rgba(149,129,255,1), rgba(130,243,236,1));
+            background: -webkit-linear-gradient(90deg, #7159c1, #bd93f9);
             -webkit-background-clip: text;
             -webkit-text-fill-color: transparent;
             color: ${props => props.theme.colors.essentials};