+++ styles/theme/ThemeDark.js
@@ -2,12 +2,12 @@ export default {
     title: 'dark',
 
     colors: {
-        essentials: '#9580ff',
-        writing: '#15171b',
-        bgcolor: '#15171b',
+        essentials: '#bd93f9',
+        writing: '#282a33',
+        bgcolor: '#282a33',
         fontcolor: '#FFF',
-        bgheader: '#1d1e26',
+        bgheader: '#282a36',
         switch: '#fff',
-        switchOn: '#9580ff',
+        switchOn: '#bd93f9',
     }
 }