+++ styles/theme/light.js
@@ -2,12 +2,12 @@ export default {
     title: 'light',
 
     colors: {
-        essentials: '#8989df',
+        essentials: '#bd93f9',
         writing: '#656565',
         bgcolor: '#fff',
         fontcolor: '#656565',
         bgheader: '#fff',
         switch: '#202020',
-        switchOn: '#8989df',
+        switchOn: '#bd93f9',
     }
 }